requirements.txt
Pillow>=10.0.0
tkcalendar>=1.6.1

